requirements.txt
streamlit>=1.28.0
yt-dlp>=2024.1.0

openai-whisper>=20231117
torch